add mapper DTOs to Models

+++ src/main/java/com/mufic/Final/api/v2/model/CourseInTermDTO.java
@@ -9,7 +9,7 @@ public class CourseInTermDTO {
 
 
     private Long id;
-    private int term;
+    private Long term;
 
     private String course;
